Line evaluated props UBO moved to layer level.
