+++ bazel/core.bzl
@@ -1072,7 +1072,6 @@ MLN_DRAWABLES_MTL_SOURCE = [
     "src/mbgl/shaders/mtl/hillshade.cpp",
     "src/mbgl/shaders/mtl/hillshade_prepare.cpp",
     "src/mbgl/shaders/mtl/line.cpp",
-    "src/mbgl/shaders/mtl/line_gradient.cpp",
     "src/mbgl/shaders/mtl/raster.cpp",
     "src/mbgl/shaders/mtl/symbol_icon.cpp",
     "src/mbgl/shaders/mtl/symbol_sdf.cpp",
@@ -1117,7 +1116,6 @@ MLN_DRAWABLES_MTL_HEADERS = [
     "include/mbgl/shaders/mtl/hillshade.hpp",
     "include/mbgl/shaders/mtl/hillshade_prepare.hpp",
     "include/mbgl/shaders/mtl/line.hpp",
-    "include/mbgl/shaders/mtl/line_gradient.hpp",
     "include/mbgl/shaders/mtl/raster.hpp",
     "include/mbgl/shaders/mtl/shader_group.hpp",
     "include/mbgl/shaders/mtl/shader_program.hpp",
