+++ include/mbgl/shaders/line_layer_ubo.hpp
@@ -15,23 +15,12 @@ struct alignas(16) LineDynamicUBO {
 };
 static_assert(sizeof(LineDynamicUBO) == 16);
 
-struct alignas(16) LineUBO {
+struct alignas(16) LineDrawableUBO {
     std::array<float, 4 * 4> matrix;
     float ratio;
     float pad1, pad2, pad3;
 };
-static_assert(sizeof(LineUBO) % 16 == 0);
-
-struct alignas(16) LinePropertiesUBO {
-    Color color;
-    float blur;
-    float opacity;
-    float gapwidth;
-    float offset;
-    float width;
-    float pad1, pad2, pad3;
-};
-static_assert(sizeof(LinePropertiesUBO) % 16 == 0);
+static_assert(sizeof(LineDrawableUBO) % 16 == 0);
 
 struct alignas(16) LineInterpolationUBO {
     float color_t;
@@ -47,17 +36,7 @@ static_assert(sizeof(LineInterpolationUBO) % 16 == 0);
 //
 // Line gradient
 
-using LineGradientUBO = LineUBO;
-
-struct alignas(16) LineGradientPropertiesUBO {
-    float blur;
-    float opacity;
-    float gapwidth;
-    float offset;
-    float width;
-    float pad1, pad2, pad3;
-};
-static_assert(sizeof(LineGradientPropertiesUBO) % 16 == 0);
+using LineGradientDrawableUBO = LineDrawableUBO;
 
 struct alignas(16) LineGradientInterpolationUBO {
     float blur_t;
@@ -72,24 +51,14 @@ static_assert(sizeof(LineGradientInterpolationUBO) % 16 == 0);
 //
 // Line pattern
 
-struct alignas(16) LinePatternUBO {
+struct alignas(16) LinePatternDrawableUBO {
     std::array<float, 4 * 4> matrix;
     std::array<float, 4> scale;
     std::array<float, 2> texsize;
     float ratio;
     float fade;
 };
-static_assert(sizeof(LinePatternUBO) % 16 == 0);
-
-struct alignas(16) LinePatternPropertiesUBO {
-    float blur;
-    float opacity;
-    float offset;
-    float gapwidth;
-    float width;
-    float pad1, pad2, pad3;
-};
-static_assert(sizeof(LinePatternPropertiesUBO) % 16 == 0);
+static_assert(sizeof(LinePatternDrawableUBO) % 16 == 0);
 
 struct alignas(16) LinePatternInterpolationUBO {
     float blur_t;
@@ -112,7 +81,7 @@ static_assert(sizeof(LinePatternTilePropertiesUBO) % 16 == 0);
 //
 // Line SDF
 
-struct alignas(16) LineSDFUBO {
+struct alignas(16) LineSDFDrawableUBO {
     std::array<float, 4 * 4> matrix;
     std::array<float, 2> patternscale_a;
     std::array<float, 2> patternscale_b;
@@ -123,19 +92,7 @@ struct alignas(16) LineSDFUBO {
     float mix;
     float pad1, pad2, pad3;
 };
-static_assert(sizeof(LineSDFUBO) % 16 == 0);
-
-struct alignas(16) LineSDFPropertiesUBO {
-    Color color;
-    float blur;
-    float opacity;
-    float gapwidth;
-    float offset;
-    float width;
-    float floorwidth;
-    float pad1, pad2;
-};
-static_assert(sizeof(LineSDFPropertiesUBO) % 16 == 0);
+static_assert(sizeof(LineSDFDrawableUBO) % 16 == 0);
 
 struct alignas(16) LineSDFInterpolationUBO {
     float color_t;
@@ -149,12 +106,27 @@ struct alignas(16) LineSDFInterpolationUBO {
 };
 static_assert(sizeof(LineSDFInterpolationUBO) % 16 == 0);
 
+//
+// Line evaluated properties
+
+struct alignas(16) LineEvaluatedPropsUBO {
+    Color color;
+    float blur;
+    float opacity;
+    float gapwidth;
+    float offset;
+    float width;
+    float floorwidth;
+    float pad1, pad2;
+};
+static_assert(sizeof(LineEvaluatedPropsUBO) % 16 == 0);
+
 enum {
     idLineDynamicUBO,
-    idLineUBO,
-    idLinePropertiesUBO,
+    idLineDrawableUBO,
     idLineInterpolationUBO,
     idLineTilePropertiesUBO,
+    idLineEvaluatedPropsUBO,
     lineUBOCount
 };
