+++ include/mbgl/shaders/mtl/line.hpp
@@ -61,11 +61,27 @@ struct FragmentStage {
 #endif
 };
 
+struct alignas(16) LineDrawableUBO {
+    float4x4 matrix;
+    float ratio;
+    float pad1, pad2, pad3;
+};
+
+struct alignas(16) LineInterpolationUBO {
+    float color_t;
+    float blur_t;
+    float opacity_t;
+    float gapwidth_t;
+    float offset_t;
+    float width_t;
+    float pad1, pad2;
+};
+
 FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
                                 device const LineDynamicUBO& dynamic [[buffer(0)]],
-                                device const LineUBO& line [[buffer(1)]],
-                                device const LinePropertiesUBO& props [[buffer(2)]],
-                                device const LineInterpolationUBO& interp [[buffer(3)]]) {
+                                device const LineDrawableUBO& drawable [[buffer(1)]],
+                                device const LineInterpolationUBO& interp [[buffer(2)]],
+                                device const LineEvaluatedPropsUBO& props [[buffer(4)]]) {
 
 #if defined(HAS_UNIFORM_u_gapwidth)
     const auto gapwidth = props.gapwidth / 2;
@@ -112,8 +128,8 @@ FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
     const float t = 1.0 - abs(u);
     const float2 offset2 = offset * a_extrude * LINE_NORMAL_SCALE * v_normal.y * float2x2(t, -u, u, t);
 
-    const float4 projected_extrude = line.matrix * float4(dist / line.ratio, 0.0, 0.0);
-    const float4 position = line.matrix * float4(pos + offset2 / line.ratio, 0.0, 1.0) + projected_extrude;
+    const float4 projected_extrude = drawable.matrix * float4(dist / drawable.ratio, 0.0, 0.0);
+    const float4 position = drawable.matrix * float4(pos + offset2 / drawable.ratio, 0.0, 1.0) + projected_extrude;
 
     // calculate how much the perspective view squishes or stretches the extrude
     const float extrude_length_without_perspective = length(dist);
@@ -138,8 +154,7 @@ FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
 }
 
 half4 fragment fragmentMain(FragmentStage in [[stage_in]],
-                            device const LineUBO& line [[buffer(1)]],
-                            device const LinePropertiesUBO& props [[buffer(2)]]) {
+                            device const LineEvaluatedPropsUBO& props [[buffer(4)]]) {
 #if defined(OVERDRAW_INSPECTOR)
     return half4(1.0);
 #endif
@@ -167,6 +182,175 @@ half4 fragment fragmentMain(FragmentStage in [[stage_in]],
     // line in case of an offset line (v_width2.y) or when fading out (v_width2.x)
     const float blur2 = (blur + 1.0 / DEVICE_PIXEL_RATIO) * in.gamma_scale;
     const float alpha = clamp(min(dist - (in.width2.y - blur2), in.width2.x - dist) / blur2, 0.0, 1.0);
+
+    return half4(color * (alpha * opacity));
+}
+)";
+};
+
+template <>
+struct ShaderSource<BuiltIn::LineGradientShader, gfx::Backend::Type::Metal> {
+    static constexpr auto name = "LineGradientShader";
+    static constexpr auto vertexMainFunction = "vertexMain";
+    static constexpr auto fragmentMainFunction = "fragmentMain";
+
+    static const std::array<UniformBlockInfo, 4> uniforms;
+    static const std::array<AttributeInfo, 7> attributes;
+    static constexpr std::array<AttributeInfo, 0> instanceAttributes{};
+    static const std::array<TextureInfo, 1> textures;
+
+    static constexpr auto source = R"(
+struct VertexStage {
+    short2 pos_normal [[attribute(5)]];
+    uchar4 data [[attribute(6)]];
+    float2 blur [[attribute(7)]];
+    float2 opacity [[attribute(8)]];
+    float2 gapwidth [[attribute(9)]];
+    float2 offset [[attribute(10)]];
+    float2 width [[attribute(11)]];
+};
+
+struct FragmentStage {
+    float4 position [[position, invariant]];
+    float2 width2;
+    half2 normal;
+    half gamma_scale;
+    float lineprogress;
+
+#if !defined(HAS_UNIFORM_u_blur)
+    half blur;
+#endif
+#if !defined(HAS_UNIFORM_u_opacity)
+    half opacity;
+#endif
+};
+
+struct alignas(16) LineGradientDrawableUBO {
+    float4x4 matrix;
+    float ratio;
+    float pad1, pad2, pad3;
+};
+
+struct alignas(16) LineGradientInterpolationUBO {
+    float blur_t;
+    float opacity_t;
+    float gapwidth_t;
+    float offset_t;
+    float width_t;
+    float pad1, pad2, pad3;
+};
+
+FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
+                                device const LineDynamicUBO& dynamic [[buffer(0)]],
+                                device const LineGradientDrawableUBO& drawable [[buffer(1)]],
+                                device const LineGradientInterpolationUBO& interp [[buffer(2)]],
+                                device const LineEvaluatedPropsUBO& props [[buffer(4)]]) {
+
+#if !defined(HAS_UNIFORM_u_blur)
+    const auto blur     = unpack_mix_float(vertx.blur,     interp.blur_t);
+#endif
+#if !defined(HAS_UNIFORM_u_opacity)
+    const auto opacity  = unpack_mix_float(vertx.opacity,  interp.opacity_t);
+#endif
+#if defined(HAS_UNIFORM_u_gapwidth)
+    const auto gapwidth = props.gapwidth;
+#else
+    const auto gapwidth = unpack_mix_float(vertx.gapwidth, interp.gapwidth_t) / 2;
+#endif
+#if defined(HAS_UNIFORM_u_offset)
+    const auto offset   = props.offset;
+#else
+    const auto offset   = unpack_mix_float(vertx.offset,   interp.offset_t) * -1;
+#endif
+#if defined(HAS_UNIFORM_u_width)
+    const auto width    = props.width;
+#else
+    const auto width    = unpack_mix_float(vertx.width,    interp.width_t);
+#endif
+
+    // the distance over which the line edge fades out.
+    // Retina devices need a smaller distance to avoid aliasing.
+    const float ANTIALIASING = 1.0 / DEVICE_PIXEL_RATIO / 2.0;
+
+    const float2 a_extrude = float2(vertx.data.xy) - 128.0;
+    const float a_direction = fmod(float(vertx.data.z), 4.0) - 1.0;
+    const float v_lineprogress = (floor(float(vertx.data.z) / 4.0) + vertx.data.w * 64.0) * 2.0 / MAX_LINE_DISTANCE;
+    const float2 pos = floor(float2(vertx.pos_normal) * 0.5);
+
+    // x is 1 if it's a round cap, 0 otherwise
+    // y is 1 if the normal points up, and -1 if it points down
+    // We store these in the least significant bit of a_pos_normal
+    const float2 normal = float2(vertx.pos_normal) - 2.0 * pos;
+    const float2 v_normal = float2(normal.x, normal.y * 2.0 - 1.0);
+
+    const float halfwidth = width / 2.0;
+    const float inset = gapwidth + (gapwidth > 0.0 ? ANTIALIASING : 0.0);
+    const float outset = gapwidth + halfwidth * (gapwidth > 0.0 ? 2.0 : 1.0) + (halfwidth == 0.0 ? 0.0 : ANTIALIASING);
+
+    // Scale the extrusion vector down to a normal and then up by the line width of this vertex.
+    const float2 dist = outset * a_extrude * LINE_NORMAL_SCALE;
+
+    // Calculate the offset when drawing a line that is to the side of the actual line.
+    // We do this by creating a vector that points towards the extrude, but rotate
+    // it when we're drawing round end points (a_direction = -1 or 1) since their
+    // extrude vector points in another direction.
+    const float u = 0.5 * a_direction;
+    const float t = 1.0 - abs(u);
+    const float2 offset2 = offset * a_extrude * LINE_NORMAL_SCALE * v_normal.y * float2x2(t, -u, u, t);
+
+    const float4 projected_extrude = drawable.matrix * float4(dist / drawable.ratio, 0.0, 0.0);
+    const float4 position = drawable.matrix * float4(pos + offset2 / drawable.ratio, 0.0, 1.0) + projected_extrude;
+
+    // calculate how much the perspective view squishes or stretches the extrude
+    const float extrude_length_without_perspective = length(dist);
+    const float extrude_length_with_perspective = length(projected_extrude.xy / position.w * dynamic.units_to_pixels);
+
+    return {
+        .position     = position,
+        .width2       = float2(outset, inset),
+        .normal       = half2(v_normal),
+        .gamma_scale  = half(extrude_length_without_perspective / extrude_length_with_perspective),
+        .lineprogress = v_lineprogress,
+
+#if !defined(HAS_UNIFORM_u_blur)
+        .blur         = half(blur),
+#endif
+#if !defined(HAS_UNIFORM_u_opacity)
+        .opacity      = half(opacity),
+#endif
+    };
+}
+
+half4 fragment fragmentMain(FragmentStage in [[stage_in]],
+                            device const LineEvaluatedPropsUBO& props [[buffer(4)]],
+                            texture2d<float, access::sample> gradientTexture [[texture(0)]]) {
+#if defined(OVERDRAW_INSPECTOR)
+    return half4(1.0);
+#endif
+
+#if defined(HAS_UNIFORM_u_blur)
+    const float blur = props.blur;
+#else
+    const float blur = float(in.blur);
+#endif
+#if defined(HAS_UNIFORM_u_opacity)
+    const float opacity = props.opacity;
+#else
+    const float opacity = float(in.opacity);
+#endif
+
+    // Calculate the distance of the pixel from the line in pixels.
+    const float dist = length(in.normal) * in.width2.x;
+
+    // Calculate the antialiasing fade factor. This is either when fading in the
+    // line in case of an offset line (v_width2.y) or when fading out (v_width2.x)
+    const float blur2 = (blur + 1.0 / DEVICE_PIXEL_RATIO) * in.gamma_scale;
+    const float alpha = clamp(min(dist - (in.width2.y - blur2), in.width2.x - dist) / blur2, 0.0, 1.0);
+
+    // For gradient lines, v_lineprogress is the ratio along the entire line,
+    // scaled to [0, 2^15), and the gradient ramp is stored in a texture.
+    constexpr sampler sampler2d(coord::normalized, filter::linear);
+    const float4 color = gradientTexture.sample(sampler2d, float2(in.lineprogress, 0.5));
 
     return half4(color * (alpha * opacity));
 }
@@ -233,21 +417,12 @@ struct FragmentStage {
 #endif
 };
 
-struct alignas(16) LinePatternUBO {
+struct alignas(16) LinePatternDrawableUBO {
     float4x4 matrix;
     float4 scale;
     float2 texsize;
     float ratio;
     float fade;
-};
-
-struct alignas(16) LinePatternPropertiesUBO {
-    float blur;
-    float opacity;
-    float offset;
-    float gapwidth;
-    float width;
-    float pad1, pad2, pad3;
 };
 
 struct alignas(16) LinePatternInterpolationUBO {
@@ -268,10 +443,10 @@ struct alignas(16) LinePatternTilePropertiesUBO {
 
 FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
                                 device const LineDynamicUBO& dynamic [[buffer(0)]],
-                                device const LinePatternUBO& line [[buffer(1)]],
-                                device const LinePatternPropertiesUBO& props [[buffer(2)]],
-                                device const LinePatternInterpolationUBO& interp [[buffer(3)]],
-                                device const LinePatternTilePropertiesUBO& tileProps [[buffer(4)]]) {
+                                device const LinePatternDrawableUBO& drawable [[buffer(1)]],
+                                device const LinePatternInterpolationUBO& interp [[buffer(2)]],
+                                device const LinePatternTilePropertiesUBO& tileProps [[buffer(3)]],
+                                device const LineEvaluatedPropsUBO& props [[buffer(4)]]) {
 
 #if defined(HAS_UNIFORM_u_gapwidth)
     const auto gapwidth = props.gapwidth / 2;
@@ -320,8 +495,8 @@ FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
     const float t = 1.0 - abs(u);
     const float2 offset2 = offset * a_extrude * LINE_NORMAL_SCALE * v_normal.y * float2x2(t, -u, u, t);
 
-    const float4 projected_extrude = line.matrix * float4(dist / line.ratio, 0.0, 0.0);
-    const float4 position = line.matrix * float4(pos + offset2 / line.ratio, 0.0, 1.0) + projected_extrude;
+    const float4 projected_extrude = drawable.matrix * float4(dist / drawable.ratio, 0.0, 0.0);
+    const float4 position = drawable.matrix * float4(pos + offset2 / drawable.ratio, 0.0, 1.0) + projected_extrude;
 
     // calculate how much the perspective view squishes or stretches the extrude
     const float extrude_length_without_perspective = length(dist);
@@ -350,9 +525,9 @@ FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
 }
 
 half4 fragment fragmentMain(FragmentStage in [[stage_in]],
-                            device const LinePatternUBO& line [[buffer(1)]],
-                            device const LinePatternPropertiesUBO& props [[buffer(2)]],
-                            device const LinePatternTilePropertiesUBO& tileProps [[buffer(4)]],
+                            device const LinePatternDrawableUBO& drawable [[buffer(1)]],
+                            device const LinePatternTilePropertiesUBO& tileProps [[buffer(3)]],
+                            device const LineEvaluatedPropsUBO& props [[buffer(4)]],
                             texture2d<float, access::sample> image0 [[texture(0)]],
                             sampler image0_sampler [[sampler(0)]]) {
 #if defined(OVERDRAW_INSPECTOR)
@@ -385,10 +560,10 @@ half4 fragment fragmentMain(FragmentStage in [[stage_in]],
     const float2 pattern_tl_b = pattern_to.xy;
     const float2 pattern_br_b = pattern_to.zw;
 
-    const float pixelRatio = line.scale.x;
-    const float tileZoomRatio = line.scale.y;
-    const float fromScale = line.scale.z;
-    const float toScale = line.scale.w;
+    const float pixelRatio = drawable.scale.x;
+    const float tileZoomRatio = drawable.scale.y;
+    const float fromScale = drawable.scale.z;
+    const float toScale = drawable.scale.w;
 
     const float2 display_size_a = float2((pattern_br_a.x - pattern_tl_a.x) / pixelRatio, (pattern_br_a.y - pattern_tl_a.y) / pixelRatio);
     const float2 display_size_b = float2((pattern_br_b.x - pattern_tl_b.x) / pixelRatio, (pattern_br_b.y - pattern_tl_b.y) / pixelRatio);
@@ -415,10 +590,10 @@ half4 fragment fragmentMain(FragmentStage in [[stage_in]],
     // the texture coordinate
     const float y_a = 0.5 + (in.normal.y * clamp(in.width2.x, 0.0, (pattern_size_a.y + 2.0) / 2.0) / pattern_size_a.y);
     const float y_b = 0.5 + (in.normal.y * clamp(in.width2.x, 0.0, (pattern_size_b.y + 2.0) / 2.0) / pattern_size_b.y);
-    const float2 pos_a = mix(pattern_tl_a / line.texsize, pattern_br_a / line.texsize, float2(x_a, y_a));
-    const float2 pos_b = mix(pattern_tl_b / line.texsize, pattern_br_b / line.texsize, float2(x_b, y_b));
+    const float2 pos_a = mix(pattern_tl_a / drawable.texsize, pattern_br_a / drawable.texsize, float2(x_a, y_a));
+    const float2 pos_b = mix(pattern_tl_b / drawable.texsize, pattern_br_b / drawable.texsize, float2(x_b, y_b));
 
-    const float4 color = mix(image0.sample(image0_sampler, pos_a), image0.sample(image0_sampler, pos_b), line.fade);
+    const float4 color = mix(image0.sample(image0_sampler, pos_a), image0.sample(image0_sampler, pos_b), drawable.fade);
 
     return half4(color * alpha * opacity);
 }
@@ -486,7 +661,7 @@ struct FragmentStage {
 #endif
 };
 
-struct alignas(16) LineSDFUBO {
+struct alignas(16) LineSDFDrawableUBO {
     float4x4 matrix;
     float2 patternscale_a;
     float2 patternscale_b;
@@ -496,17 +671,6 @@ struct alignas(16) LineSDFUBO {
     float sdfgamma;
     float mix;
     float pad1, pad2, pad3;
-};
-
-struct alignas(16) LineSDFPropertiesUBO {
-    float4 color;
-    float blur;
-    float opacity;
-    float gapwidth;
-    float offset;
-    float width;
-    float floorwidth;
-    float pad1, pad2;
 };
 
 struct alignas(16) LineSDFInterpolationUBO {
@@ -522,9 +686,9 @@ struct alignas(16) LineSDFInterpolationUBO {
 
 FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
                                 device const LineDynamicUBO& dynamic [[buffer(0)]],
-                                device const LineSDFUBO& line [[buffer(1)]],
-                                device const LineSDFPropertiesUBO& props [[buffer(2)]],
-                                device const LineSDFInterpolationUBO& interp [[buffer(3)]]) {
+                                device const LineSDFDrawableUBO& drawable [[buffer(1)]],
+                                device const LineSDFInterpolationUBO& interp [[buffer(2)]],
+                                device const LineEvaluatedPropsUBO& props [[buffer(4)]]) {
 
 #if defined(HAS_UNIFORM_u_gapwidth)
     const auto gapwidth   = props.gapwidth / 2;
@@ -578,8 +742,8 @@ FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
     const float t = 1.0 - abs(u);
     const float2 offset2 = offset * a_extrude * LINE_NORMAL_SCALE * v_normal.y * float2x2(t, -u, u, t);
 
-    const float4 projected_extrude = line.matrix * float4(dist / line.ratio, 0.0, 0.0);
-    const float4 position = line.matrix * float4(pos + offset2 / line.ratio, 0.0, 1.0) + projected_extrude;
+    const float4 projected_extrude = drawable.matrix * float4(dist / drawable.ratio, 0.0, 0.0);
+    const float4 position = drawable.matrix * float4(pos + offset2 / drawable.ratio, 0.0, 1.0) + projected_extrude;
 
     // calculate how much the perspective view squishes or stretches the extrude
     const float extrude_length_without_perspective = length(dist);
@@ -590,8 +754,8 @@ FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
         .width2       = float2(outset, inset),
         .normal       = v_normal,
         .gamma_scale  = half(extrude_length_without_perspective / extrude_length_with_perspective),
-        .tex_a        = float2(linesofar * line.patternscale_a.x / floorwidth, (normal.y * line.patternscale_a.y + line.tex_y_a) * 2.0),
-        .tex_b        = float2(linesofar * line.patternscale_b.x / floorwidth, (normal.y * line.patternscale_b.y + line.tex_y_b) * 2.0),
+        .tex_a        = float2(linesofar * drawable.patternscale_a.x / floorwidth, (normal.y * drawable.patternscale_a.y + drawable.tex_y_a) * 2.0),
+        .tex_b        = float2(linesofar * drawable.patternscale_b.x / floorwidth, (normal.y * drawable.patternscale_b.y + drawable.tex_y_b) * 2.0),
 
 #if !defined(HAS_UNIFORM_u_color)
         .color        = unpack_mix_color(vertx.color, interp.color_t),
@@ -609,8 +773,8 @@ FragmentStage vertex vertexMain(thread const VertexStage vertx [[stage_in]],
 }
 
 half4 fragment fragmentMain(FragmentStage in [[stage_in]],
-                            device const LineSDFUBO& line [[buffer(1)]],
-                            device const LineSDFPropertiesUBO& props [[buffer(2)]],
+                            device const LineSDFDrawableUBO& drawable [[buffer(1)]],
+                            device const LineEvaluatedPropsUBO& props [[buffer(4)]],
                             texture2d<float, access::sample> image0 [[texture(0)]],
                             sampler image0_sampler [[sampler(0)]]) {
 #if defined(OVERDRAW_INSPECTOR)
@@ -644,10 +808,10 @@ half4 fragment fragmentMain(FragmentStage in [[stage_in]],
 
     const float sdfdist_a = image0.sample(image0_sampler, in.tex_a).a;
     const float sdfdist_b = image0.sample(image0_sampler, in.tex_b).a;
-    const float sdfdist = mix(sdfdist_a, sdfdist_b, line.mix);
+    const float sdfdist = mix(sdfdist_a, sdfdist_b, drawable.mix);
     const float dist = length(in.normal) * in.width2.x;
     const float alpha = clamp(min(dist - (in.width2.y - blur2), in.width2.x - dist) / blur2, 0.0, 1.0) *
-                        smoothstep(0.5 - line.sdfgamma / floorwidth, 0.5 + line.sdfgamma / floorwidth, sdfdist);
+                        smoothstep(0.5 - drawable.sdfgamma / floorwidth, 0.5 + drawable.sdfgamma / floorwidth, sdfdist);
 
     return half4(color * (alpha * opacity));
 }
