--- src/mbgl/shaders/mtl/line_gradient.cpp
@@ -1,26 +1,0 @@
-#include <mbgl/shaders/mtl/line_gradient.hpp>
-
-namespace mbgl {
-namespace shaders {
-
-const std::array<UniformBlockInfo, 4> ShaderSource<BuiltIn::LineGradientShader, gfx::Backend::Type::Metal>::uniforms = {
-    UniformBlockInfo{true, false, sizeof(LineDynamicUBO), idLineDynamicUBO},
-    UniformBlockInfo{true, true, sizeof(LineGradientUBO), idLineUBO},
-    UniformBlockInfo{true, true, sizeof(LineGradientPropertiesUBO), idLinePropertiesUBO},
-    UniformBlockInfo{true, false, sizeof(LineGradientInterpolationUBO), idLineInterpolationUBO},
-};
-const std::array<AttributeInfo, 7> ShaderSource<BuiltIn::LineGradientShader, gfx::Backend::Type::Metal>::attributes = {
-    AttributeInfo{lineUBOCount + 0, gfx::AttributeDataType::Short2, idLinePosNormalVertexAttribute},
-    AttributeInfo{lineUBOCount + 1, gfx::AttributeDataType::UByte4, idLineDataVertexAttribute},
-    AttributeInfo{lineUBOCount + 2, gfx::AttributeDataType::Float2, idLineBlurVertexAttribute},
-    AttributeInfo{lineUBOCount + 3, gfx::AttributeDataType::Float2, idLineOpacityVertexAttribute},
-    AttributeInfo{lineUBOCount + 4, gfx::AttributeDataType::Float2, idLineGapWidthVertexAttribute},
-    AttributeInfo{lineUBOCount + 5, gfx::AttributeDataType::Float2, idLineOffsetVertexAttribute},
-    AttributeInfo{lineUBOCount + 6, gfx::AttributeDataType::Float2, idLineWidthVertexAttribute},
-};
-const std::array<TextureInfo, 1> ShaderSource<BuiltIn::LineGradientShader, gfx::Backend::Type::Metal>::textures = {
-    TextureInfo{0, idLineImageTexture},
-};
-
-} // namespace shaders
-} // namespace mbgl
